# [Dev] Improve typescript performance and version bump node and dependencies