--- a/src/@types/common.ts
+++ b/src/@types/common.ts
@@ -1,1 +1,4 @@
 export type ConditionFn = (args?: any[]) => boolean;
+
+/** A union type of all primitives (types that are always passed by value) */
+export type Primitive = string | number | boolean | bigint | null | undefined | symbol;