--- a/src/@types/pokemon-level-moves.ts
+++ b/src/@types/pokemon-level-moves.ts
@@ -1,0 +1,19 @@
+/*
+ * SPDX-FileCopyrightText: 2025 Pagefault Games
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ */
+
+import type { MoveId } from "#enums/move-id";
+
+export type LevelMoves = [number, MoveId][];
+
+export interface PokemonSpeciesLevelMoves {
+  [key: number]: LevelMoves;
+}
+interface PokemonFormLevelMoves {
+  [key: number]: LevelMoves;
+}
+export interface PokemonSpeciesFormLevelMoves {
+  [key: number]: PokemonFormLevelMoves;
+}