--- a/src/data/balance/pokemon-level-moves.ts
+++ b/src/data/balance/pokemon-level-moves.ts
@@ -3,6 +3,7 @@
  * @internal
  * @module
  */
+
 import { EVOLVE_MOVE, RELEARN_MOVE } from "#app/constants";
 import { MoveId } from "#enums/move-id";
 import { SpeciesId } from "#enums/species-id";