--- a/src/data/data-lists.ts
+++ b/src/data/data-lists.ts
@@ -1,7 +1,9 @@
 import type { Ability } from "#abilities/ability";
 import type { PokemonSpecies } from "#data/pokemon-species";
+import type { SpeciesId } from "#enums/species-id";
 import type { ModifierTypes } from "#modifiers/modifier-type";
 import type { Move } from "#moves/move";
+import type { BiomeDepths, CatchableSpecies } from "#types/biomes";
 
 export const allAbilities: Ability[] = [];
 export const allMoves: Move[] = [];
@@ -9,3 +11,7 @@ export const allSpecies: PokemonSpecies[] = [];
 
 // TODO: Figure out what this is used for and provide an appropriate tsdoc comment
 export const modifierTypes = {} as ModifierTypes;
+
+export const catchableSpecies: CatchableSpecies = {};
+export const biomeDepths: BiomeDepths = {};
+export const uncatchableSpecies: SpeciesId[] = [];