--- a/src/data/mystery-encounters/utils/encounter-phase-utils.ts
+++ b/src/data/mystery-encounters/utils/encounter-phase-utils.ts
@@ -58,7 +58,7 @@ import i18next from "i18next";
  * Animates exclamation sprite over trainer's head at start of encounter
  * @param scene
  */
-export function doTrainerExclamation() {
+export function doTrainerExclamation(): void {
   const exclamationSprite = globalScene.add.sprite(0, 0, "encounter_exclaim");
   exclamationSprite.setName("exclamation");
   globalScene.field.add(exclamationSprite);
@@ -447,9 +447,9 @@ export async function initBattleWithEnemyConfig(partyConfig: EnemyPartyConfig): 
  * This promise does not need to be awaited on if called in an encounter onInit (will just load lazily)
  * @param moves
  */
-export function loadCustomMovesForEncounter(moves: MoveId | MoveId[]) {
-  moves = coerceArray(moves);
-  return Promise.all(moves.map(move => initMoveAnim(move))).then(() => loadMoveAnimAssets(moves));
+export async function loadCustomMovesForEncounter(moves: MoveId | MoveId[]): Promise<void> {
+  const movesArray: MoveId[] = coerceArray(moves);
+  return Promise.all(movesArray.map((move: MoveId) => initMoveAnim(move))).then(() => loadMoveAnimAssets(movesArray));
 }
 
 /**
@@ -458,7 +458,7 @@ export function loadCustomMovesForEncounter(moves: MoveId | MoveId[]) {
  * @param playSound
  * @param showMessage
  */
-export function updatePlayerMoney(changeValue: number, playSound = true, showMessage = true) {
+export function updatePlayerMoney(changeValue: number, playSound = true, showMessage = true): void {
   globalScene.money = Math.min(Math.max(globalScene.money + changeValue, 0), Number.MAX_SAFE_INTEGER);
   globalScene.updateMoneyText();
   globalScene.animateMoneyChanged(false);
@@ -730,7 +730,7 @@ export function setEncounterRewards(
   customShopRewards?: CustomModifierSettings,
   eggRewards?: IEggOptions[],
   preRewardsCallback?: Function,
-) {
+): void {
   globalScene.currentBattle.mysteryEncounter!.doEncounterRewards = () => {
     if (preRewardsCallback) {
       preRewardsCallback();
@@ -798,7 +798,7 @@ export class OptionSelectSettings {
  * MUST be used only in onOptionPhase, will not work in onPreOptionPhase or onPostOptionPhase
  * @param optionSelectSettings
  */
-export function initSubsequentOptionSelect(optionSelectSettings: OptionSelectSettings) {
+export function initSubsequentOptionSelect(optionSelectSettings: OptionSelectSettings): void {
   globalScene.phaseManager.pushNew("MysteryEncounterPhase", optionSelectSettings);
 }
 
@@ -811,7 +811,7 @@ export function initSubsequentOptionSelect(optionSelectSettings: OptionSelectSet
 export function leaveEncounterWithoutBattle(
   addHealPhase = false,
   encounterMode: MysteryEncounterMode = MysteryEncounterMode.NO_BATTLE,
-) {
+): void {
   globalScene.currentBattle.mysteryEncounter!.encounterMode = encounterMode;
   globalScene.phaseManager.clearPhaseQueue(true);
   handleMysteryEncounterVictory(addHealPhase);
@@ -822,7 +822,7 @@ export function leaveEncounterWithoutBattle(
  * @param addHealPhase - Adds an empty shop phase to allow player to purchase healing items
  * @param doNotContinue - default `false`. If set to true, will not end the battle and continue to next wave
  */
-export function handleMysteryEncounterVictory(addHealPhase = false, doNotContinue = false) {
+export function handleMysteryEncounterVictory(addHealPhase = false, doNotContinue = false): void {
   const allowedPkm = globalScene.getPlayerParty().filter(pkm => pkm.isAllowedInBattle());
 
   if (allowedPkm.length === 0) {
@@ -865,7 +865,7 @@ export function handleMysteryEncounterVictory(addHealPhase = false, doNotContinu
  * Similar to {@linkcode handleMysteryEncounterVictory}, but for cases where the player lost a battle or failed a challenge
  * @param addHealPhase
  */
-export function handleMysteryEncounterBattleFailed(addHealPhase = false, doNotContinue = false) {
+export function handleMysteryEncounterBattleFailed(addHealPhase = false, doNotContinue = false): void {
   const allowedPkm = globalScene.getPlayerParty().filter(pkm => pkm.isAllowedInBattle());
 
   if (allowedPkm.length === 0) {
@@ -945,7 +945,7 @@ export function transitionMysteryEncounterIntroVisuals(hide = true, destroy = tr
  * Will queue moves for any pokemon to use before the first CommandPhase of a battle
  * Mostly useful for allowing {@linkcode MysteryEncounter} enemies to "cheat" and use moves before the first turn
  */
-export function handleMysteryEncounterBattleStartEffects() {
+export function handleMysteryEncounterBattleStartEffects(): void {
   const encounter = globalScene.currentBattle.mysteryEncounter;
   if (
     globalScene.currentBattle.isBattleMysteryEncounter()
@@ -1036,7 +1036,7 @@ export function getRandomEncounterSpecies(level: number, isBoss = false, rerollH
  * Just a helper function to calculate aggregate stats for MEs in a Classic run
  * @param baseSpawnWeight
  */
-export function calculateMEAggregateStats(baseSpawnWeight: number) {
+export function calculateMEAggregateStats(baseSpawnWeight: number): void {
   const numRuns = 1000;
   let run = 0;
   const biomes = Object.keys(BiomeId).filter(key => Number.isNaN(Number(key)));
@@ -1219,7 +1219,7 @@ export function calculateMEAggregateStats(baseSpawnWeight: number) {
  * Just a helper function to calculate aggregate stats for MEs in a Classic run
  * @param luckValue - 0 to 14
  */
-export function calculateRareSpawnAggregateStats(luckValue: number) {
+export function calculateRareSpawnAggregateStats(luckValue: number): void {
   const numRuns = 1000;
   let run = 0;