--- a/src/init/init.ts
+++ b/src/init/init.ts
@@ -1,5 +1,5 @@
 import { initAbilities } from "#abilities/ability";
-import { initBiomes } from "#balance/biomes";
+import { initBiomes } from "#balance/init-biomes";
 import { initPokemonPrevolutions, initPokemonStarters } from "#balance/pokemon-evolutions";
 import { initSpecies } from "#balance/pokemon-species";
 import { initChallenges } from "#data/challenge";