--- a/src/phases/evolution-phase.ts
+++ b/src/phases/evolution-phase.ts
@@ -1,10 +1,10 @@
 import type { AnySound } from "#app/battle-scene";
+import { EVOLVE_MOVE } from "#app/constants";
 import { globalScene } from "#app/global-scene";
 import { getPokemonNameWithAffix } from "#app/messages";
 import { Phase } from "#app/phase";
 import type { SpeciesFormEvolution } from "#balance/pokemon-evolutions";
 import { FusionSpeciesFormEvolution } from "#balance/pokemon-evolutions";
-import { EVOLVE_MOVE } from "#balance/pokemon-level-moves";
 import { getTypeRgb } from "#data/type";
 import { LearnMoveSituation } from "#enums/learn-move-situation";
 import { UiMode } from "#enums/ui-mode";