--- a/test/data/balance/pokemon-level-moves.test.ts
+++ b/test/data/balance/pokemon-level-moves.test.ts
@@ -14,6 +14,7 @@
  *
  * @module
  */
+
 import { pokemonFormLevelMoves, pokemonSpeciesLevelMoves } from "#balance/pokemon-level-moves";
 import { MoveId } from "#enums/move-id";
 import { SpeciesId } from "#enums/species-id";