> [!IMPORTANT]
> You must update node to use this PR.
 
## What are the changes the user will see?
N/A

## Why am I making these changes?
TypeScript is taking too long to parse certain files and it is causing slowdowns when using it.

## What are the changes from a developer perspective?

- Since it is October, Node 24 is now LTS, so we are upgrading 
- Extracted out `init-biomes` from `src/data/balance/biomes.ts`
- Added `ts-nocheck` to `pokemon-level-moves` and extracted types out of it to its own file
- Moved tmSpecies to its own file that has a `ts-nocheck` directive at the top
- Added type annotations to methods in `data/mystery-encounters/utils/encounter-phase-utils.ts`
- Updated each package that was outdated by using `pnpm outadated`, after sufficiently investigating potential breaking changes and compromised package updates.

This is *purely* a code movement PR. There are *no* changes to code logic.

#### Discussion
So that the massive several thousand-line structures were placed in their own file. For `tms` and `pokemon-level-moves`, added an ``ts-nocheck`` directive to the top to make typecheck skip type checking this file. To remedy this, I created tests for these files to ensure that types work as expected, so that we do not sacrifice any type safety by using no check.
Also, using a type assertion `via as ...` instead of a type annotation is faster for some reason, so I place it here.

## Comparison

<details><summary>Stats (before)</summary>

❯ pnpm exec tsc --extendedDiagnostics --noEmit
Files:                         2133
Lines of Library:             52164
Lines of Definitions:        262461
Lines of TypeScript:         329695
Lines of JavaScript:              0
Lines of JSON:                 2922
Lines of Other:                   0
Identifiers:                 876852
Symbols:                     964725
Types:                       421556
Instantiations:              964475
Memory used:               1546440K
Assignability cache size:    469823
Identity cache size:          20265
Subtype cache size:           24045
Strict subtype cache size:    27881
I/O Read time:                0.10s
Parse time:                   1.51s
ResolveModule time:           0.17s
ResolveTypeReference time:    0.00s
ResolveLibrary time:          0.01s
Program time:                 1.95s
Bind time:                    0.65s
Check time:                  21.11s
printTime time:               0.00s
Emit time:                    0.00s
Total time:                  23.71s

</details>

<details><summary>Stats (after)</summary>

❯ pnpm exec tsc --extendedDiagnostics --noEmit
Files:                         2151
Lines of Library:             52804
Lines of Definitions:        264577
Lines of TypeScript:         329879
Lines of JavaScript:              0
Lines of JSON:                 2922
Lines of Other:                   0
Identifiers:                 880295
Symbols:                     769286
Types:                       177362
Instantiations:              581790
Memory used:               1104332K
Assignability cache size:     90726
Identity cache size:           4719
Subtype cache size:           24253
Strict subtype cache size:    27935
I/O Read time:                0.10s
Parse time:                   1.52s
ResolveModule time:           0.18s
ResolveTypeReference time:    0.00s
ResolveLibrary time:          0.01s
Program time:                 1.96s
Bind time:                    0.65s
Check time:                  11.36s
printTime time:               0.00s
Emit time:                    0.00s
Total time:                  13.98s
</details> 

#### Summary

|                 | **Time**      | **Memory**  |
|-----------------|---------------|-------------|
| **Before**      | 23.71 seconds | ~1510.20 MB |
| **After**       | 13.98 seconds | ~1078.45 MB |
| **Improvement** | 41.03%        | 28.59%      |

So we can see we save a substantial amount of time and memory. I expect the time difference to be even greater on slower machines.

## How to test the changes?

First, update node (hopefully you're using nvm, if you're not, well, get node 24.9.0 and `pnpm`

The reason to update node first is to avoid polluting the timing results with any differences from the node version increase, which wouldn't really be fair.

```sh
pnpm self-update
nvm install .
nvm use 24
```

To run a comparison, use
```sh
pnpm exec tsc --extendedDiagnostics --noEmit
```
on beta BEFORE checking out this PR.

Then, switch to this branch, update node, and run the command again.


## Checklist
- [x] **I'm using `beta` as my base branch**
- [x] There is no overlap with another PR?
- [x] The PR is self-contained and cannot be split into smaller PRs?
- [x] Have I provided a clear explanation of the changes?
- [x] Have I tested the changes manually?
- [x] Are all unit tests still passing? (`pnpm test:silent`)
  - [x] Have I created new automated tests (`pnpm test:create`) or updated existing tests related to the PR's changes?
- ~~[ ] Have I provided screenshots/videos of the changes (if applicable)?~~
  - ~~[ ] Have I made sure that any UI change works for both UI themes (default and legacy)?~~